[PON-7] Re-build default layout, fix product interface
