--- src/components/patials/header/const.ts
@@ -1,12 +1,0 @@
-const linksList = [
-  {
-    href: '/number',
-    title: 'Random number'
-  },
-  {
-    href: '/wheel',
-    title: 'Random wheel'
-  }
-];
-
-export {linksList};
